Javascript error handling (fixed all errors in js,html) v0.9

+++ src/html/script.js
@@ -6,32 +6,12 @@ $(document).ready(function () {
     initAddBtn();
     initCancelBtn();
     initSaveBtn();
-    initAddRoleBtn();
+    initSaveRoleButton();
+    initAddRoleButton();
+    fetchAllz();
+    document.getElementById("persons").disable = true;
 
 });
-
-function initAddRoleBtn() {
-    $("#btn_role").click(function () {
-        initRoleDetails(true);
-        //First create post argument as a JavaScript object
-        var newSchoolRole = {"degree": $("#pfunction").val(), "roleName": $("#prole").val()};
-        //{"degree": "BITCH","roleName": "Teacher"}
-        alert(JSON.stringify(newSchoolRole));
-        $.ajax({
-            url: "../role/" + $("#id").val(),
-            data: JSON.stringify(newSchoolRole), //Convert newRole to JSON
-            type: "post",
-            dataType: 'json',
-            error: function (jqXHR, textStatus, errorThrown) {
-                alert("Error in add new SchoolRole" + jqXHR.responseText + ": " + textStatus);
-            }
-        }).done(function (newSchoolRole) {
-           $("#id").val(newSchoolRole.roleName);
-
-        });
-    });
-}
-
 
 function initAddBtn() {
     $("#btn_add").click(function () {
@@ -40,23 +20,75 @@ function initAddBtn() {
     });
 }
 
+function initAddRoleButton() {
+    $("#btn_addrole").click(function () {
+        document.getElementById("persons").disable = true;
+        initRoleDetails(true);
+        //fetchAllz();
+    });
+}
+
+function initSaveRoleButton() {
+    $("#btn_role").click(function () {
+        
+            var newSchoolRole;
+            if ($("#prole").val() === "Teacher" && $("#prole").val() !== "" && $("#pfunction").val() !== "") {
+                newSchoolRole = {"degree": $("#pfunction").val(), "roleName": $("#prole").val()};
+                executeRoleSavingModule(newSchoolRole);
+            } else if ($("#prole").val() === "Student" && $("#prole").val() !== "" && $("#pfunction").val() !== "") {
+                newSchoolRole = {"semester": $("#pfunction").val(), "roleName": $("#prole").val()};
+                executeRoleSavingModule(newSchoolRole);
+            } else if ($("#prole").val() === "TeacherAssistant" && $("#prole").val() !== "") {
+                newSchoolRole = {"roleName": $("#prole").val()};
+                executeRoleSavingModule(newSchoolRole);
+            } else {
+                alert("PLEASE INPUT SOMETHING PROPER OR YOU WILL CRASH OUR PROGRAM </3 ... :/");
+            }
+        
+        initRoleDetails(false);
+        fetchAll();
+    });
+}
+
+function executeRoleSavingModule(newSchoolRole) {
+    alert(JSON.stringify(newSchoolRole));
+    $.ajax({
+        url: "../role/" + $("#id").val(),
+        data: JSON.stringify(newSchoolRole), //Convert newRole to JSON
+        type: "post",
+        dataType: 'json',
+        error: function (jqXHR, textStatus, errorThrown) {
+            alert("Error in add new SchoolRole" + jqXHR.responseText + ": " + textStatus);
+        }
+    });
+}
+
 function initSaveBtn() {
     $("#btn_save").click(function () {
-        //First create post argument as a JavaScript object
-        var newPerson = {"firstName": $("#fname").val(), "lastName": $("#lname").val(), "phone": $("#phone").val(), "mail": $("#email").val()};
-        $.ajax({
-            url: "../person",
-            data: JSON.stringify(newPerson), //Convert newPerson to JSON
-            type: "post",
-            dataType: 'json',
-            error: function (jqXHR, textStatus, errorThrown) {
-                alert(jqXHR.responseText + ": " + textStatus);
-            }
-        }).done(function (newPerson) {
-            $("#id").val(newPerson.id);
-            initDetails(false);
-            fetchAll();
-        });
+        var letters = /^[a-z]+$/;
+        //if(inputtxt.value.match(letters))
+
+        if (!$("#phone").val().match(letters) && $("#fname").val() !== ""
+                && $("#lname").val() !== "" && $("#phone").val() !== ""
+                && $("#email").val() !== "" && $("#email").val().match("@")) {
+            //First create post argument as a JavaScript object
+            var newPerson = {"firstName": $("#fname").val(), "lastName": $("#lname").val(), "phone": $("#phone").val(), "mail": $("#email").val()};
+            $.ajax({
+                url: "../person",
+                data: JSON.stringify(newPerson), //Convert newPerson to JSON
+                type: "post",
+                dataType: 'json',
+                error: function (jqXHR, textStatus, errorThrown) {
+                    alert(jqXHR.responseText + ": " + textStatus);
+                }
+            }).done(function (newPerson) {
+                $("#id").val(newPerson.id);
+                initDetails(false);
+                fetchAll();
+            });
+        } else {
+            alert("STOP RUINING MY CODE  ... :/");
+        }
     });
 }
 
@@ -81,16 +113,23 @@ function initPersons() {
 
 function deletePerson() {
     $("#delete").click(function () {
-        $.ajax({
-            url: "../person/" + $("#persons option:selected").attr("id"),
-            type: "DELETE",
-            dataType: 'json',
-            error: function (jqXHR, textStatus, errorThrown) {
-                alert(jqXHR.responseText + ": " + textStatus);
-            }
-        }).done(function () {
-            fetchAll();
-        });
+        if ($("#persons option:selected").val()) {
+            $.ajax({
+                url: "../person/" + $("#persons option:selected").attr("id"),
+                type: "DELETE",
+                dataType: 'json',
+                error: function (jqXHR, textStatus, errorThrown) {
+                    alert(jqXHR.responseText + ": " + textStatus);
+                    if ($("#persons option:selected").attr("id") === null) {
+                        alert("Got ya.");
+                    }
+                }
+            }).done(function () {
+                fetchAll();
+            });
+        } else {
+            alert("NOTHING SELECTED, DON'T TRY TO RUIN OUR PROGRAM PLS ;S::S:S:S;S;s;sSssS");
+        }
     });
 }
 
@@ -103,7 +142,9 @@ function initDetails(init) {
         $("#btn_save").removeAttr("disabled");
         $("#btn_cancel").removeAttr("disabled");
         $("#btn_add").attr("disabled", "disabled");
-
+        $("#btn_role").attr("disabled", "disabled");
+        $("#btn_addrole").attr("disabled", "disabled");
+        $("#delete").attr("disabled", "disabled");
 
     }
     else {
@@ -113,8 +154,10 @@ function initDetails(init) {
         $("#email").attr("disabled", "disabled");
         $("#btn_save").attr("disabled", "disabled");
         $("#btn_cancel").attr("disabled", "disabled");
+        $("#btn_role").attr("disabled", "disabled");
+        $("#btn_addrole").attr("disabled", "disabled");
         $("#btn_add").removeAttr("disabled");
-        $("#btn_role").removeAttr("disabled");
+        $("#delete").removeAttr("disabled");
     }
 }
 
@@ -122,17 +165,20 @@ function initRoleDetails(init) {
     if (init) {
         $("#prole").removeAttr("disabled");
         $("#pfunction").removeAttr("disabled");
-        $("#btn_save").removeAttr("disabled");
         $("#btn_cancel").removeAttr("disabled");
+        $("#btn_role").removeAttr("disabled");
         $("#btn_add").attr("disabled", "disabled");
+        $("#btn_addrole").attr("disabled", "disabled");
+        $("#delete").attr("disabled", "disabled");
     }
     else {
-        $("#fname").attr("disabled", "disabled");
-        $("#lname").attr("disabled", "disabled");
-        $("#phone").attr("disabled", "dsiabled");
-        $("#btn_save").attr("disabled", "disabled");
+        $("#prole").attr("disabled", "disabled");
+        $("#pfunction").attr("disabled", "disabled");
         $("#btn_cancel").attr("disabled", "disabled");
+        $("#btn_role").attr("disabled", "disabled");
+        $("#btn_addrole").attr("disabled", "disabled");
         $("#btn_add").removeAttr("disabled");
+        $("#delete").removeAttr("disabled");
     }
 }
 
@@ -142,6 +188,8 @@ function clearDetails() {
     $("#lname").val("");
     $("#phone").val("");
     $("#email").val("");
+    $("#prole").val("");
+    $("#pfunction").val("");
 }
 function updateDetails(id) {
     $.ajax({
@@ -157,7 +205,7 @@ function updateDetails(id) {
         $("#lname").val(person.lastName);
         $("#phone").val(person.phone);
         $("#email").val(person.mail);
-        $("#btn_role").removeAttr("disabled");
+        $("#btn_addrole").removeAttr("disabled");
     });
 }
 
@@ -176,5 +224,23 @@ function fetchAll() {
         });
         $("#persons").html(options);
         clearDetails();
+    });
+}
+
+function fetchAllz() {
+    $.ajax({
+        url: "../person",
+        type: "GET",
+        dataType: 'json',
+        error: function (jqXHR, textStatus, errorThrown) {
+            alert(textStatus);
+        }
+    }).done(function (persons) {
+        var options = "";
+        persons.forEach(function (person) {
+            options += "<option id=" + person.id + ">" + person.firstName + ", " + person.lastName + ", " + person.phone + ", " + person.mail + ", " + JSON.stringify(person.listOfSchoolRoles) + "</option>";
+        });
+        $("#persons").html(options);
+        //clearDetails();
     });
 }
